[UPD] mrp(workorder): setting production_id field as `required=False`
because in module `rep_sequar/sq_mrp_workorders_group` we need that
field as nullable.

We overrided it in `sq_mrp_workorders_group/models/mrp_workorder.py`,
but in every update we receive an error due to odoo updating firs the
base model, then our customization.
It's a non-blocking error, but isn't a good thing to see in update logs.